apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Api necessary packages"
authors = ["jacobadev <junior1232131@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
pytest = "^8.1.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
